requirements.txt
gradio
python-dotenv
browser-use
playwright
langchain
langchain-google-genai
pyperclip
pytest
google.genai